# docs/apifox-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "基于区块链的志愿者服务认证系统 API",
    "description": "Apifox 导入文件。非公共接口统一使用 Authorization: Bearer <token>。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发环境"
    }
  ],
  "security": [{"bearerAuth": []}],
...
  "paths": {
...
    "/api/activities": {
      "post": {
        "tags": ["活动"],
        "summary": "组织创建活动草稿",
        "requestBody": {"$ref": "#/components/requestBodies/ActivitySaveRequest"},
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
    "/api/activities/{id}": {
      "put": {
        "tags": ["活动"],
        "summary": "组织更新活动",
        "description": "已审核活动修改关键字段后会重新进入 pending_review。",
        "parameters": [{"$ref": "#/components/parameters/IdPath"}],
        "requestBody": {"$ref": "#/components/requestBodies/ActivitySaveRequest"},
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      },
      "delete": {
        "tags": ["活动"],
        "summary": "组织删除活动",
        "parameters": [{"$ref": "#/components/parameters/IdPath"}],
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
    "/api/activities/{id}/submit": {
      "post": {
        "tags": ["活动"],
        "summary": "提交活动发布审核",
        "parameters": [{"$ref": "#/components/parameters/IdPath"}],
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
...
    "/api/participations/{activityId}": {
      "post": {
        "tags": ["活动"],
        "summary": "报名活动",
        "description": "志愿者可报名任意组织已审核活动。",
        "parameters": [{"$ref": "#/components/parameters/ActivityIdPath"}],
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      },
      "delete": {
        "tags": ["活动"],
        "summary": "取消报名",
        "parameters": [{"$ref": "#/components/parameters/ActivityIdPath"}],
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
...
    "/api/completions": {
      "post": {
        "tags": ["活动"],
        "summary": "提交完成报告",
        "requestBody": {"$ref": "#/components/requestBodies/CompletionSubmitRequest"},
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
...
    "/api/reviews": {
      "get": {
        "tags": ["活动"],
        "summary": "组织完成报告审核列表",
        "parameters": [{"$ref": "#/components/parameters/StatusQuery"}],
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
    "/api/reviews/{completionId}": {
      "post": {
        "tags": ["活动"],
        "summary": "审核完成报告",
        "description": "审核通过后按 approvedRewardPoints 发放积分，并创建存证索引。",
        "parameters": [{"$ref": "#/components/parameters/CompletionIdPath"}],
        "requestBody": {"$ref": "#/components/requestBodies/CompletionReviewRequest"},
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
...
    "/api/admin/users/{userId}/roles": {
      "get": {
        "tags": ["平台审核"],
        "summary": "用户角色",
        "parameters": [{"$ref": "#/components/parameters/UserIdPath"}],
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      },
      "put": {
        "tags": ["平台审核"],
        "summary": "分配用户角色",
        "parameters": [{"$ref": "#/components/parameters/UserIdPath"}],
        "requestBody": {"$ref": "#/components/requestBodies/AssignUserRolesRequest"},
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
...
    "/api/admin/roles/{roleCode}/permissions": {
      "get": {
        "tags": ["平台审核"],
        "summary": "角色权限",
        "parameters": [{"name": "roleCode", "in": "path", "required": true, "schema": {"type": "string"}, "example": "system_admin"}],
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
...
    "/api/admin/activity-reviews": {
      "get": {
        "tags": ["平台审核"],
        "summary": "活动发布审核列表",
        "parameters": [{"$ref": "#/components/parameters/StatusQuery"}],
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
    "/api/admin/activity-reviews/{activityId}": {
      "post": {
        "tags": ["平台审核"],
        "summary": "处理活动发布审核",
        "description": "普通活动支持 approved/rejected；高风险活动先 escalated，再由 system_admin 复审。",
        "parameters": [{"$ref": "#/components/parameters/ActivityIdPath"}],
        "requestBody": {"$ref": "#/components/requestBodies/ActivityReviewRequest"},
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
    "/api/admin/product-reviews": {
      "get": {
        "tags": ["平台审核"],
        "summary": "商品审核列表",
        "parameters": [{"$ref": "#/components/parameters/StatusQuery"}],
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
    "/api/admin/product-reviews/{productId}": {
      "post": {
        "tags": ["平台审核"],
        "summary": "处理商品审核",
        "parameters": [{"$ref": "#/components/parameters/ProductIdPath"}],
        "requestBody": {"$ref": "#/components/requestBodies/ProductReviewRequest"},
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
    "/api/admin/points-rules": {
      "get": {
        "tags": ["平台审核"],
        "summary": "积分规则列表",
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      },
      "post": {
        "tags": ["平台审核"],
        "summary": "创建积分规则",
        "requestBody": {"$ref": "#/components/requestBodies/PointsRuleSaveRequest"},
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
    "/api/admin/points-rules/{id}": {
      "put": {
        "tags": ["平台审核"],
        "summary": "更新积分规则",
        "parameters": [{"$ref": "#/components/parameters/IdPath"}],
        "requestBody": {"$ref": "#/components/requestBodies/PointsRuleSaveRequest"},
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
...
    "/api/evidences/{bizType}/{bizId}": {
      "get": {
        "tags": ["存证"],
        "summary": "登录用户查询单条存证",
        "parameters": [{"$ref": "#/components/parameters/BizTypePath"}, {"$ref": "#/components/parameters/BizIdPath"}],
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
...
    "/api/evidences/{id}/retry": {
      "post": {
        "tags": ["存证"],
        "summary": "重试上链",
        "parameters": [{"$ref": "#/components/parameters/IdPath"}],
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
...
    "/api/public/evidences/{bizType}/{bizId}": {
      "get": {
        "tags": ["公共"],
        "summary": "公开存证详情",
        "security": [],
        "parameters": [{"$ref": "#/components/parameters/BizTypePath"}, {"$ref": "#/components/parameters/BizIdPath"}],
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}
    },
    "parameters": {
      "IdPath": {"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}},
      "UserIdPath": {"name": "userId", "in": "path", "required": true, "schema": {"type": "integer"}},
      "ActivityIdPath": {"name": "activityId", "in": "path", "required": true, "schema": {"type": "integer"}},
      "ProductIdPath": {"name": "productId", "in": "path", "required": true, "schema": {"type": "integer"}},
      "CompletionIdPath": {"name": "completionId", "in": "path", "required": true, "schema": {"type": "integer"}},
      "BizTypePath": {"name": "bizType", "in": "path", "required": true, "schema": {"type": "string", "enum": ["completion"]}},
      "BizIdPath": {"name": "bizId", "in": "path", "required": true, "schema": {"type": "integer"}},
      "StatusQuery": {"name": "status", "in": "query", "required": false, "schema": {"type": "string"}}
    },
    "responses": {
      "ApiResponse": {
        "description": "统一响应",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ApiResponse"},
            "example": {"success": true, "message": "操作成功", "data": {}}
          }
        }
      }
    },
    "requestBodies": {
...
      "ActivitySaveRequest": {
        "required": true,
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ActivitySaveRequest"}}}
      },
      "CompletionSubmitRequest": {
        "required": true,
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CompletionSubmitRequest"}}}
      },
      "CompletionReviewRequest": {
        "required": true,
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CompletionReviewRequest"}, "example": {"status": "approved", "reviewNote": "服务记录真实，审核通过。"}}}
      },
...
      "AssignUserRolesRequest": {
        "required": true,
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AssignUserRolesRequest"}, "example": {"roleCodes": ["activity_reviewer"]}}}
      },
      "ActivityReviewRequest": {
        "required": true,
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ActivityReviewRequest"}, "example": {"status": "approved", "approvedRewardPoints": 30, "reviewNote": "活动信息真实，积分合理。"}}}
      },
      "ProductReviewRequest": {
        "required": true,
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductReviewRequest"}, "example": {"status": "approved", "reviewNote": "商品信息完整，可以进入统一商城。"}}}
      },
      "PointsRuleSaveRequest": {
        "required": true,
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PointsRuleSaveRequest"}}}
      },
...
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean"},
          "message": {"type": "string"},
          "data": {
            "description": "实际类型随接口变化，字段明细见接口联调用例和后端 VO。",
            "nullable": true,
            "oneOf": [
              {"type": "object"},
              {"type": "array", "items": {}},
              {"type": "string"},
              {"type": "number"},
              {"type": "integer"},
              {"type": "boolean"}
            ]
          }
        }
      },
...
      "ActivitySaveRequest": {
        "type": "object",
        "required": ["activityName", "startDate", "endDate", "location", "maxParticipants", "requestedRewardPoints"],
        "properties": {
          "activityName": {"type": "string", "maxLength": 255},
          "description": {"type": "string", "maxLength": 3000},
          "startDate": {"type": "string", "example": "2026-05-10 09:00:00"},
          "endDate": {"type": "string", "example": "2026-05-10 12:00:00"},
          "location": {"type": "string", "maxLength": 255},
          "contactName": {"type": "string", "maxLength": 255},
          "contactPhone": {"type": "string", "maxLength": 20},
          "categoryTags": {"type": "string", "maxLength": 255},
          "imagePath": {"type": "string", "maxLength": 255},
          "maxParticipants": {"type": "integer", "minimum": 1},
          "requestedRewardPoints": {"type": "integer", "minimum": 0, "description": "组织申请积分"},
          "enrollDeadline": {"type": "string", "example": "2026-05-09 18:00:00"}
        },
        "example": {
          "activityName": "社区环保清洁活动",
          "description": "清理社区公共区域垃圾并进行环保宣传。",
          "startDate": "2026-05-10 09:00:00",
          "endDate": "2026-05-10 12:00:00",
          "location": "广州市天河区社区广场",
          "contactName": "李老师",
          "contactPhone": "13800000000",
          "categoryTags": "GENERAL,环保",
          "imagePath": "/uploads/projects/clean.png",
          "maxParticipants": 30,
          "requestedRewardPoints": 30,
          "enrollDeadline": "2026-05-09 18:00:00"
        }
      },
      "CompletionSubmitRequest": {
        "type": "object",
        "required": ["activityId", "serviceLocation", "serviceStartTime", "serviceEndTime", "reportText", "contributionDetails"],
        "properties": {
          "activityId": {"type": "integer"},
          "serviceLocation": {"type": "string"},
          "serviceStartTime": {"type": "string", "example": "2026-05-10 09:00:00"},
          "serviceEndTime": {"type": "string", "example": "2026-05-10 12:00:00"},
          "reportText": {"type": "string", "maxLength": 3000},
          "contributionDetails": {"type": "string", "maxLength": 3000},
          "attachmentPaths": {"type": "array", "items": {"type": "string"}}
        },
        "example": {
          "activityId": 1,
          "serviceLocation": "广州市天河区社区广场",
          "serviceStartTime": "2026-05-10 09:00:00",
          "serviceEndTime": "2026-05-10 12:00:00",
          "reportText": "完成社区道路和绿化带清洁。",
          "contributionDetails": "清理垃圾 8 袋，协助完成环保宣传。",
          "attachmentPaths": ["/uploads/projects/report-1.png"]
        }
      },
      "CompletionReviewRequest": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": {"type": "string", "enum": ["approved", "rejected"]},
          "reviewNote": {"type": "string", "maxLength": 500}
        }
      },
...
      "AssignUserRolesRequest": {
        "type": "object",
        "required": ["roleCodes"],
        "properties": {
          "roleCodes": {
            "type": "array",
            "items": {"type": "string", "enum": ["volunteer", "organization_admin", "activity_reviewer", "product_reviewer", "system_admin"]}
          }
        }
      },
      "ActivityReviewRequest": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": {"type": "string", "enum": ["approved", "rejected", "escalated"]},
          "approvedRewardPoints": {"type": "integer", "minimum": 0, "description": "审核通过时必填，积分发放读取该值。"},
          "reviewNote": {"type": "string", "maxLength": 500}
        }
      },
      "ProductReviewRequest": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": {"type": "string", "enum": ["approved", "rejected", "off_shelf"]},
          "reviewNote": {"type": "string", "maxLength": 500}
        }
      },
      "PointsRuleSaveRequest": {
        "type": "object",
        "required": ["activityCategory", "minServiceHours", "maxServiceHours", "suggestedPoints", "maxPoints", "escalationThreshold", "status"],
        "properties": {
          "activityCategory": {"type": "string", "maxLength": 100},
          "minServiceHours": {"type": "number"},
          "maxServiceHours": {"type": "number"},
          "suggestedPoints": {"type": "integer", "minimum": 0},
          "maxPoints": {"type": "integer", "minimum": 0},
          "escalationThreshold": {"type": "integer", "minimum": 0},
          "status": {"type": "integer", "enum": [0, 1]}
        },
        "example": {
          "activityCategory": "GENERAL",
          "minServiceHours": 0,
          "maxServiceHours": 4,
          "suggestedPoints": 30,
          "maxPoints": 50,
          "escalationThreshold": 80,
          "status": 1
        }
      }
    }
  }
}